Atualizado o retorno para a contagem no Status.

+++ src/main/java/br/com/meli/simiandna/vo/StatsVO.java
@@ -16,13 +16,13 @@ import lombok.ToString;
 @ToString
 public class StatsVO {
 
-    private double count_mutant_dna;
-    private double count_human_dna;
+    private int count_mutant_dna;
+    private int count_human_dna;
     
     public double getRatio() {
     	
-    	if(count_human_dna <= 0 || count_mutant_dna <= 0) return 0;
+    	if(count_human_dna <= 0 || count_mutant_dna <= 0) return 0d;
 		
-		return new BigDecimal(count_mutant_dna / count_human_dna).setScale(6,RoundingMode.HALF_UP).doubleValue();
+		return new BigDecimal((double)count_mutant_dna / (double)count_human_dna).setScale(6,RoundingMode.HALF_UP).doubleValue();
     }
 }
